Properly seperate A7 support from the LX45 and further work on DDR3 iface
I've generated a new DDR3 ip using the files provided for the Arty-A7-35T, which
is the target board for this port.  Additionally, I have redone the top level
pins to match the DDR3 interface (lots of renames, a few addtions and removals).

The design still is not completely routed, but it is much closer -- only two
circuits are unroutable.  I will continue investigating this.