Fix SDHC initialization, I think.
Sadly it's not easy to tell until the entire thing boots because I don't have
a simulation model of an SDHC card, but it does appear to get all the way to
the idle state.  The problem is that it seems to either stay there, or it
could be that the VGA display doesn't actually work.