Add LED controller for debugging, and try to fix SDHC and timing.
The LED controller permits me to make much more effective use of the Arty's
LEDs, including moded operation (different switch settings --> different
meanings for each LED).

The largest timing fix was to introduce a synchronizer for the global reset,
which vastly improved timing.  Most of the remaining "error" is on the other
reset lines, but it's not neccessary to "fix" them (I did it to reduce the
synthesis/implementation time).  The ideal "fix" would be to tell Vivado
that these are multi-cycle paths, but when I attempted both that and also
setting them as false-paths, Vivado then decided that said paths no longer
existed, and threw an error about it.

I also added an outside SD controller for a future attempt to replace the
current controller.